update swearwords.go
i edited some words because i dont think 'shit' is not such a BIG badword. so i changed it into false. and some word had spelling mistakes like before it was 'fvuk' but i corrected it to 'fuck' the correct one in slot number 72. BUT ALL OTHER WORDS ARE LIKE WHEN IT WAS BEFORE. I DIDNT EDIT!! thank uuu... I ALSO WANT MY SERVER TO BE MORE SAFE!!!